Move pkg/queue to internal/queue
